requirements.txt
flask==3.0.3
pyjwt==2.9.0
cryptography==43.0.3
pytest==8.3.3
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.9